requirements.txt
django
django-unfold
factory-boy
